refactor: colocate blank stack navigation helpers
