+++ packages/react-native-screen-transitions/src/shared/providers/stack/blank-stack-state/helpers/navigation/compose-descriptors.ts
@@ -1,4 +1,4 @@
-import type { RouteWithKey } from "../../types/stack.types";
+import type { RouteWithKey } from "../../../../../types/stack.types";
 
 export const composeDescriptors = <
 	Route extends RouteWithKey,
