+++ packages/react-native-screen-transitions/src/shared/providers/stack/blank-stack-state/helpers/navigation/resolve-scene-neighbors.ts
