+++ packages/react-native-screen-transitions/src/shared/providers/stack/blank-stack-state/helpers/navigation/sync-routes-with-removed.ts
@@ -1,4 +1,4 @@
-import type { RouteWithKey } from "../../types/stack.types";
+import type { RouteWithKey } from "../../../../../types/stack.types";
 import { composeDescriptors } from "./compose-descriptors";
 
 type SyncRoutesWithRemovedParams<
