+++ packages/react-native-screen-transitions/src/shared/providers/stack/blank-stack-state/helpers/reconcile-blank-stack-routes.ts
@@ -2,8 +2,8 @@ import type {
 	BaseStackDescriptor,
 	RouteWithKey,
 } from "../../../../types/stack.types";
-import { composeDescriptors } from "../../../../utils/navigation/compose-descriptors";
-import { syncRoutesWithRemoved } from "../../../../utils/navigation/sync-routes-with-removed";
+import { composeDescriptors } from "./navigation/compose-descriptors";
+import { syncRoutesWithRemoved } from "./navigation/sync-routes-with-removed";
 import { routesHaveSameKeys } from "./state-equality";
 import type {
 	BlankStackDescriptorSources,
